Modified JsonFormatter: now it's for right before HTML rendering
